+++ config/config.exs
@@ -2,12 +2,10 @@ use Mix.Config
 
 if Mix.env() == :bench do
   config :ua_inspector,
-    database_path: Path.expand("../data", __DIR__),
-    startup_sync: true
+    database_path: Path.expand("../data", __DIR__)
 end
 
 if Mix.env() == :test do
   config :ua_inspector,
-    database_path: Path.expand("../test/fixtures", __DIR__),
-    startup_sync: true
+    database_path: Path.expand("../test/fixtures", __DIR__)
 end