+++ lib/ua_inspector/config.ex
@@ -22,7 +22,7 @@ defmodule UAInspector.Config do
         ],
         remote_release: "master",
         startup_silent: false,
-        startup_sync: false,
+        startup_sync: true,
         yaml_file_reader: {:yamerl_constr, :file, [[:str_node_as_binary]]}
 
   The default `:database_path` is evaluated at runtime and not compiled into
@@ -75,15 +75,14 @@ defmodule UAInspector.Config do
 
   ## Startup Behaviour
 
-  Databases are loaded asynchronously when starting the application using a
-  `:reload` cast. This default behaviour can lead to the first parsing calls to
-  work with an empty database and therefore not return the results you expect.
-
-  You can change this behaviour to have the application force a synchronous
+  You can change this behaviour to have the application force an asynchronous
   database loading during the initial startup:
 
       config :ua_inspector,
-        startup_sync: true
+        startup_sync: false
+
+  This can lead to the first parsing calls to work with an empty database
+  and therefore not return the results you expect.
 
   ### Starting Silently