+++ lib/ua_inspector/storage/server.ex
@@ -13,7 +13,7 @@ defmodule UAInspector.Storage.Server do
       @ets_table_opts [:named_table, :protected, :set, read_concurrency: true]
 
       def init(_init_arg) do
-        if Config.get(:startup_sync, false) do
+        if Config.get(:startup_sync, true) do
           :ok = reload_database()
         else
           :ok = GenServer.cast(__MODULE__, :reload)