+++ verify/config/config.exs
@@ -1,5 +1,4 @@
 use Mix.Config
 
 config :ua_inspector,
-  database_path: Path.expand("../databases", __DIR__),
-  startup_sync: true
+  database_path: Path.expand("../databases", __DIR__)