+++ CHANGELOG.md
@@ -7,6 +7,7 @@
         - `UAInspector.Downloader.read_remote/0`
         - `UAInspector.Downloader.README.path_local/0`
         - `UAInspector.Downloader.README.path_priv/0`
+    - Startup is now done with a blocking database load by default
 
 ## v1.2.0 (2019-08-10)